Test: MockRandomNumberGenerator 재사용을 위한 분리
