+++ src/test/java/racingcar/domain/car/CarsTest.java
@@ -2,11 +2,9 @@ package racingcar.domain.car;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.random.RandomNumberGenerator;
+import racingcar.domain.mock.MockRandomNumberGenerator;
 
-import java.util.LinkedList;
 import java.util.List;
-import java.util.Queue;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.*;
@@ -96,30 +94,5 @@ class CarsTest {
                 () -> assertEquals(2, topCarNames.size()),
                 () -> assertEquals(names, topCarNames)
         );
-    }
-
-    static class MockRandomNumberGenerator implements RandomNumberGenerator {
-
-        private final Queue<Integer> numbers;
-
-        private MockRandomNumberGenerator(Queue<Integer> numbers) {
-            this.numbers = numbers;
-        }
-
-        public static RandomNumberGenerator setNumbers(int... numbers) {
-            Queue<Integer> queue = new LinkedList<>();
-            for (int number : numbers) {
-                queue.add(number);
-            }
-            return new MockRandomNumberGenerator(queue);
-        }
-
-        @Override
-        public int getRandomNumber() {
-            if (!numbers.isEmpty()) {
-                return numbers.remove();
-            }
-            throw new IllegalArgumentException("숫자가 더이상 존재하지 않음");
-        }
     }
 }
